+++ src/authentication/authentication.tsx
@@ -1,20 +1,9 @@
 import { z } from "zod";
-import React, {
-  createContext,
-  useState,
-  useEffect,
-  useMemo,
-  useCallback
-} from "react";
 import * as Msal from "@azure/msal-browser";
 import { jwtDecode } from "jwt-decode";
 import PublicApi from "../api/public";
 import { OrganizationsDataApi } from "../api/generated";
 import {
-  authenticationStateSchema,
-  CurrentSession,
-  UserSession,
-  empty,
   deleteUserNonceByState,
   getUserNonceByState,
   setUserNonceByState,
@@ -26,31 +15,48 @@ import {
   getAdminNonceByState
 } from "./authenticationState";
 import { makeStore } from "./authenticationStore";
-import { load, save, watch } from "./authenticationPersistance";
+import {
+  localStorageInitialState,
+  localStorageSetup
+} from "./authenticationPersistance";
 
 // TODO remove tokens and nonces that are expired
 // TODO replace /session?... on page load
 // TODO delete nonce and session on login error and show error message
 
-const LOCAL_STORAGE_KEY = "oneidentity";
+export const authenticationStore = makeStore(localStorageInitialState);
+localStorageSetup(authenticationStore);
 
-export const authenticationStore = makeStore(
-  load({
-    key: LOCAL_STORAGE_KEY,
-    validate: authenticationStateSchema.parse,
-    empty
-  })
+function randomAlphaNumericString(length: number): string {
+  const choices =
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+  const buffer = new Uint8Array(length);
+  crypto.getRandomValues(buffer);
+  return Array.from(buffer, byte => choices[byte % choices.length]).join("");
+}
+
+const userJWTPayloadSchema = z.object({
+  role: z.literal("ROLE_MERCHANT"),
+  fiscal_code: z.string(),
+  first_name: z.string(),
+  last_name: z.string(),
+  iat: z.number(),
+  exp: z.number()
+});
+
+const adminJWTPayloadSchema = z.object({
+  role: z.literal("ROLE_ADMIN"),
+  first_name: z.string(),
+  last_name: z.string(),
+  iat: z.number(),
+  exp: z.number()
+});
+
+const organizationsDataApi = new OrganizationsDataApi(
+  undefined,
+  process.env.BASE_API_PATH,
+  undefined
 );
-
-authenticationStore.subscribe(() => {
-  save({ key: LOCAL_STORAGE_KEY, value: authenticationStore.get() });
-});
-
-watch({
-  key: LOCAL_STORAGE_KEY,
-  validate: authenticationStateSchema.parse,
-  listener: authenticationStore.set
-});
 
 export function goToUserLoginPage() {
   const targetHost = process.env.ONE_IDENTITY_LOGIN_HOST ?? "";
@@ -72,83 +78,7 @@ export function goToUserLoginPage() {
   window.location.href = targetUrl;
 }
 
-const AdminAccess = new Msal.PublicClientApplication({
-  auth: {
-    clientId: process.env.MSAL_CLIENT_ID as string,
-    authority: process.env.MSAL_AUTHORITY as string,
-    knownAuthorities: [
-      "testcgnportalbitrock.b2clogin.com",
-      "cgnonboardingportaluat.b2clogin.com",
-      "cgnonboardingportal.b2clogin.com"
-    ], // You must identify your tenant's domain as a known authority.
-    redirectUri: process.env.MSAL_REDIRECT_URI as string,
-    postLogoutRedirectUri: "/"
-  },
-  cache: {
-    cacheLocation: "sessionStorage",
-    storeAuthStateInCookie: false
-  }
-});
-
-export function goToAdminLoginPage() {
-  const state = randomAlphaNumericString(16);
-  const nonce = randomAlphaNumericString(16);
-  setAdminNonceByState(state, nonce);
-  void AdminAccess.loginRedirect({
-    scopes: ["openid", process.env.MSAL_CLIENT_ID as string],
-    state,
-    nonce
-  });
-}
-
-const userJWTPayloadSchema = z.object({
-  role: z.literal("ROLE_MERCHANT"),
-  fiscal_code: z.string(),
-  first_name: z.string(),
-  last_name: z.string(),
-  iat: z.number(),
-  exp: z.number()
-});
-
-const adminJWTPayloadSchema = z.object({
-  role: z.literal("ROLE_ADMIN"),
-  first_name: z.string(),
-  last_name: z.string(),
-  iat: z.number(),
-  exp: z.number()
-});
-
-void (async () => {
-  const result = await AdminAccess.handleRedirectPromise();
-  if (result) {
-    const nonce = getAdminNonceByState(result.state ?? "");
-    if (nonce) {
-      deleteAdminNonceByState(result.state ?? "");
-      const adminTokenResponse = await PublicApi.Session.createJwtSessionToken({
-        createJwtSessionTokenRequest: {
-          requestType: "ad",
-          token: result.idToken,
-          nonce
-        }
-      });
-      const decoded = jwtDecode<unknown>(adminTokenResponse.data);
-      // eslint-disable-next-line
-      console.log(decoded);
-      const parsed = adminJWTPayloadSchema.parse(decoded);
-      // TODO salvare admin token e impostare current session
-    }
-  }
-})();
-
-function randomAlphaNumericString(length: number): string {
-  const choices =
-    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-  const buffer = new Uint8Array(length);
-  crypto.getRandomValues(buffer);
-  return Array.from(buffer, byte => choices[byte % choices.length]).join("");
-}
-
-async function oneIdentityLoginCallback() {
+void (async function onUserLoginRedirect() {
   const url = new URL(window.location.href);
   if (url.pathname !== "/session") {
     return;
@@ -180,15 +110,58 @@ async function oneIdentityLoginCallback() {
   setCurrentSession({ type: "user", userFiscalCode: fiscal_code });
   // eslint-disable-next-line functional/immutable-data
   window.location.href = "/";
+})();
+
+const AdminAccess = new Msal.PublicClientApplication({
+  auth: {
+    clientId: process.env.MSAL_CLIENT_ID as string,
+    authority: process.env.MSAL_AUTHORITY as string,
+    knownAuthorities: [
+      "testcgnportalbitrock.b2clogin.com",
+      "cgnonboardingportaluat.b2clogin.com",
+      "cgnonboardingportal.b2clogin.com"
+    ], // You must identify your tenant's domain as a known authority.
+    redirectUri: process.env.MSAL_REDIRECT_URI as string,
+    postLogoutRedirectUri: "/"
+  },
+  cache: {
+    cacheLocation: "sessionStorage",
+    storeAuthStateInCookie: false
+  }
+});
+
+export function goToAdminLoginPage() {
+  const state = randomAlphaNumericString(16);
+  const nonce = randomAlphaNumericString(16);
+  setAdminNonceByState(state, nonce);
+  void AdminAccess.loginRedirect({
+    scopes: ["openid", process.env.MSAL_CLIENT_ID as string],
+    state,
+    nonce
+  });
 }
 
-setTimeout(oneIdentityLoginCallback, 0);
-
-const organizationsDataApi = new OrganizationsDataApi(
-  undefined,
-  process.env.BASE_API_PATH,
-  undefined
-);
+void (async function onAdminLoginRedirect() {
+  const result = await AdminAccess.handleRedirectPromise();
+  if (result) {
+    const nonce = getAdminNonceByState(result.state ?? "");
+    if (nonce) {
+      deleteAdminNonceByState(result.state ?? "");
+      const adminTokenResponse = await PublicApi.Session.createJwtSessionToken({
+        createJwtSessionTokenRequest: {
+          requestType: "ad",
+          token: result.idToken,
+          nonce
+        }
+      });
+      const decoded = jwtDecode<unknown>(adminTokenResponse.data);
+      // eslint-disable-next-line
+      console.log(decoded);
+      const parsed = adminJWTPayloadSchema.parse(decoded);
+      // TODO salvare admin token e impostare current session
+    }
+  }
+})();
 
 export function on401() {
   const data = authenticationStore.get();
@@ -196,62 +169,4 @@ export function on401() {
   setCurrentSession(null);
   // eslint-disable-next-line functional/immutable-data
   window.location.href = "/";
-}
-
-export type AuthenticationContextType = {
-  userSessionByFiscalCode: Record<string, UserSession>;
-  currentSession: CurrentSession;
-  changeSession(session: CurrentSession): void;
-  logout(session: CurrentSession): void;
-};
-
-const AuthenticationContext = createContext<AuthenticationContextType>(
-  null as any
-);
-
-export function AuthenticationProvider({
-  children
-}: {
-  children: React.ReactNode;
-}) {
-  const [persistedState, setPersistedState] = useState(authenticationStore.get);
-  useEffect(
-    () =>
-      authenticationStore.subscribe(() => {
-        setPersistedState(authenticationStore.get());
-      }),
-    []
-  );
-  const { currentSession, userSessionByFiscalCode } = persistedState;
-  const changeSession = useCallback((session: CurrentSession) => {
-    setCurrentSession(session);
-    // eslint-disable-next-line functional/immutable-data
-    window.location.href = "/";
-  }, []);
-  const logout = useCallback((session: CurrentSession) => {
-    deleteSession(session);
-    setCurrentSession(null);
-    // eslint-disable-next-line functional/immutable-data
-    window.location.href = "/";
-  }, []);
-  const value = useMemo<AuthenticationContextType>(
-    () => ({
-      userSessionByFiscalCode,
-      currentSession,
-      changeSession,
-      logout
-    }),
-    [changeSession, currentSession, logout, userSessionByFiscalCode]
-  );
-  return (
-    <AuthenticationContext.Provider value={value}>
-      {children}
-    </AuthenticationContext.Provider>
-  );
-}
-
-export const AuthenticationConsumer = AuthenticationContext.Consumer;
-
-export function useAuthentication() {
-  return React.useContext(AuthenticationContext);
 }
